fit: auth, jwt, hash. Service-struct encapsulation

+++ pkg/models/movie.go
@@ -1,7 +1,7 @@
 package models
 
 type Movie struct {
-	ID uint `json:"id" gorm:"primary_key"`
-	Title string `json:"title"`
+	ID     int64  `json:"id" gorm:"primary_key"`
+	Title  string `json:"title"`
 	Author string `json:"author"`
 }
